Mv version and status test into runtime_info.go
